+++ config-typesafe-kodein/src/main/kotlin/uy/klutter/config/typesafe/kodein/ConfigInjection.kt
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.databind.type.TypeFactory
 import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
 import com.github.salomonbrys.kodein.Kodein
+import com.github.salomonbrys.kodein.instance
 import com.github.salomonbrys.kodein.singleton
 import com.github.salomonbrys.kodein.typeToken
 import com.typesafe.config.Config
@@ -15,8 +16,10 @@ fun Kodein.Builder.importConfig(config: Config, configExpression: String? = null
     import(KodeinTypesafeConfig.Module(init).Builder(root, mapper).kodeinModule, allowOverride)
 }
 
+// fun Kodein.Companion.ConfigModule(init: KodeinTypesafeConfig.Module.Builder.()->Unit) = KodeinTypesafeConfig.Module { init() }
+
 class KodeinTypesafeConfig private constructor() {
-    class Module(private val init: Builder.(Config) -> Unit) {
+    class Module internal constructor (private val init: Builder.(Config) -> Unit) {
         inner class Builder(val config: Config, val mapper: ObjectMapper = jacksonObjectMapper()) {
             internal val kodeinModule = Kodein.Module() {
                 // call init which builds a list of things to do
@@ -57,8 +60,9 @@ class KodeinTypesafeConfig private constructor() {
                 private fun addAction(targetConfig: Config) {
                     actions.add(BindActions {
                         val asJson = targetConfig.root().render(ConfigRenderOptions.concise().setJson(true))
-                        val instance: T = mapper.readValue(asJson, TypeFactory.defaultInstance().constructType(_bind.type))!!
-                        DirectBinder(_bind.tag, overrides)._with(_bind.type, singleton { instance })
+                        val value: T = mapper.readValue(asJson, TypeFactory.defaultInstance().constructType(_bind.type))!!
+  //                      bind(_bind.type, _bind.tag) with instance(value)
+                        DirectBinder(_bind.tag, overrides)._with(_bind.type, singleton { value })
                     })
                 }
             }
@@ -76,6 +80,7 @@ class KodeinTypesafeConfig private constructor() {
                 private fun addAction(configValue: ConfigValue) {
                     actions.add(BindActions {
                         val value = configValue.unwrapped() as T
+ //  constant(_bind.tag!!, overrides) with value
                         this.ConstantBinder(_bind.tag!!, overrides) with(value)
                     })
                 }